Make "+ Add book" button functional with dialog
The button is not functional. by pressing on the button, a
modal box will pop up, in which user will fill a form and submit it.
The information about the book and whether it's been read or not is saved
and will be shown on a card on the screen.